+++ sdmx-dl-provider-ri/src/test/java/internal/sdmxdl/ri/web/monitors/UpptimeMonitoringTest.java
@@ -19,21 +19,30 @@ public class UpptimeMonitoringTest {
     public void testSiteSummary() throws IOException {
         try (Reader reader = new InputStreamReader(Objects.requireNonNull(UpptimeMonitoringTest.class.getResourceAsStream("summary.json")), StandardCharsets.UTF_8)) {
             assertThat(UpptimeMonitoring.SiteSummary.parseAll(reader))
-                    .contains(UpptimeMonitoring.SiteSummary.of("ABS", "up", "100.00%"), atIndex(0))
-                    .contains(UpptimeMonitoring.SiteSummary.of("ILO", "down", "20.97%"), atIndex(1))
+                    .contains(of("ABS", "up", "100.00%", 4674), atIndex(0))
+                    .contains(of("ILO", "down", "20.97%", 14989), atIndex(1))
                     .hasSize(2);
         }
     }
 
     @Test
     public void testGetReport() {
-        assertThat(UpptimeMonitoring.getReport(UpptimeMonitoring.SiteSummary.of("ABS", "up", "100.00%")))
-                .isEqualTo(SdmxWebMonitorReport.builder().status(SdmxWebStatus.UP).uptimeRatio(1d).build());
+        assertThat(UpptimeMonitoring.getReport(of("ABS", "up", "100.00%", 4674)))
+                .isEqualTo(SdmxWebMonitorReport.builder().status(SdmxWebStatus.UP).uptimeRatio(1d).averageResponseTime(4674L).build());
 
-        assertThat(UpptimeMonitoring.getReport(UpptimeMonitoring.SiteSummary.of("ILO", "down", "20.97%")))
-                .isEqualTo(SdmxWebMonitorReport.builder().status(SdmxWebStatus.DOWN).uptimeRatio(.2097d).build());
+        assertThat(UpptimeMonitoring.getReport(of("ILO", "down", "20.97%", 14989)))
+                .isEqualTo(SdmxWebMonitorReport.builder().status(SdmxWebStatus.DOWN).uptimeRatio(.2097d).averageResponseTime(14989L).build());
 
-        assertThat(UpptimeMonitoring.getReport(UpptimeMonitoring.SiteSummary.of("ILO", "down", "z")))
-                .isEqualTo(SdmxWebMonitorReport.builder().status(SdmxWebStatus.DOWN).uptimeRatio(0d).build());
+        assertThat(UpptimeMonitoring.getReport(of("ILO", "down", "X", -1)))
+                .isEqualTo(SdmxWebMonitorReport.builder().status(SdmxWebStatus.DOWN).averageResponseTime(-1L).build());
+    }
+
+    static UpptimeMonitoring.SiteSummary of(String name, String status, String uptime, long time) {
+        UpptimeMonitoring.SiteSummary result = new UpptimeMonitoring.SiteSummary();
+        result.setName(name);
+        result.setStatus(status);
+        result.setUptime(uptime);
+        result.setTime(time);
+        return result;
     }
 }